requirements.txt
Django==5.0.3
whitenoise==6.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg[binary]
scikit-learn>=1.5.0
joblib==1.3.2
pandas>=2.2.0
numpy>=1.26.0
drf-yasg==1.21.7
gunicorn==21.2.0
python-dotenv==1.0.1
setuptools<70.0.0
django-filter==24.1
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.3
Pillow==10.3.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.2.1
bcrypt==4.1.3
pyotp==2.9.0
django-cryptography==1.1
requests==2.31.0
